strangercollective/requirements.txt
altgraph==0.17
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.8.2
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
colorama==0.4.3
cryptography==36.0.1
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
Django==3.0.9
django-admin-search==0.3.5
django-background-tasks==1.2.5
hpack==3.0.0
django-colorfield==0.3.2
django-compat==1.0.15
django-cors-headers==3.4.0
django-filter==2.3.0
django-markdownx==3.0.1
django-modelcluster==5.0.2
django-tagging==0.5.0
django-taggit==1.3.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-csv==2.1.0
docutils==0.16
draftjs-exporter==2.1.7
emails==0.6
et-xmlfile==1.0.1
future==0.18.2
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.27
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==1.7.0
inflect==0.2.5
install==1.3.3
isort==4.3.21
jdcal==1.4.1
jeepney==0.7.1
joblib==0.17.0
keyring==21.5.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
llvmlite==0.36.0
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.4.0
mccabe==0.6.1